IsWindowBorderless api function. Set editor area in text editor, work in progress.

+++ Code/Berta/Source/Berta/Controls/TextEditors/TextEditorBase.h
@@ -58,7 +58,7 @@ namespace Berta
 		size_t logicalLineIndex;
 		size_t charStart;
 		size_t charLength;
-		float y; // Posición vertical absoluta
+		uint32_t y; // Posición vertical absoluta
 	};
 }
